@@ -7,7 +7,6 @@ import (
 	"flag"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"log"
 	"net"
 	"net/http"
@@ -21,6 +20,8 @@ import (
 	"strings"
 	"sync"
 	"time"
+
+	"github.com/xeipuuv/gojsonschema"
 )
 
 var (
@@ -47,10 +48,11 @@ var (
 
 // TestCase defines a single case to be tested
 type TestCase struct {
-	Name string `json:"name"`
-	Err  string `json:"error"`
-	In   Input  `json:"in"`
-	Out  Output `json:"out"`
+	Name string     `json:"name"`
+	Err  string     `json:"error"`
+	In   Input      `json:"in"`
+	Out  Output     `json:"out"`
+	Next []TestCase `json:"next"`
 }
 
 // Input is the definition of the request to send in a given TestCase
@@ -63,9 +65,10 @@ type Input struct {
 
 // Output contains the data required to verify the response received in a given TestCase
 type Output struct {
-	StatusCode int                 `json:"status_code"`
-	Body       interface{}         `json:"body"`
-	Header     map[string][]string `json:"header"`
+	StatusCode int                    `json:"status_code"`
+	Body       interface{}            `json:"body"`
+	Header     map[string][]string    `json:"header"`
+	Schema     map[string]interface{} `json:"schema"`
 }
 
 // CmdBuilder defines an interface for building the cmd to be managed by the Runner
@@ -76,6 +79,20 @@ type CmdBuilder interface {
 // BackendBuilder defines an interface for building a server as a backend for the tests
 type BackendBuilder interface {
 	New(*Config) http.Server
+}
+
+// GenericServer defines an interface to launch a server that
+// could be an http.Server, a different type, or a wrapper
+// around multiple servers.
+type GenericServer interface {
+	Close() error
+	ListenAndServe() error
+}
+
+// ComposableBackendBuilder allows us to return
+// a more generic interface for any kind of server.
+type ComposableBackendBuilder interface {
+	NewGenericServer(*Config) GenericServer
 }
 
 // Config contains options for running a test.
@@ -184,10 +201,18 @@ func NewIntegration(cfg *Config, cb CmdBuilder, bb BackendBuilder) (*Runner, []T
 	}
 
 	if bb == nil {
-		bb = defaultBackendBuilder
+		bb = DefaultBackendBuilder
 	}
 
-	backend := bb.New(cfg)
+	var backend GenericServer
+	cbb, ok := bb.(ComposableBackendBuilder)
+	if ok {
+		backend = cbb.NewGenericServer(cfg)
+	} else {
+		httpServer := bb.New(cfg)
+		backend = &httpServer
+	}
+
 	closeFuncs = append(closeFuncs, func() { backend.Close() })
 
 	go func() {
@@ -275,9 +300,9 @@ func assertResponse(actual *http.Response, expected Output) error {
 	}
 
 	var body interface{}
-
+	var bodyBytes []byte
 	if actual.Body != nil {
-		b, err := ioutil.ReadAll(actual.Body)
+		b, err := io.ReadAll(actual.Body)
 		if err != nil {
 			return err
 		}
@@ -289,10 +314,42 @@ func assertResponse(actual *http.Response, expected Output) error {
 		default:
 			_ = json.Unmarshal(b, &body)
 		}
+		bodyBytes = b
 	}
 
-	if !reflect.DeepEqual(body, expected.Body) {
-		errMsgs = append(errMsgs, fmt.Sprintf("unexpected body.\n\t\thave: %v\n\t\twant: %v", body, expected.Body))
+	if len(expected.Schema) != 0 {
+		if len(bodyBytes) == 0 {
+			return responseError{
+				errMessage: append(errMsgs, "cannot validate empty body"),
+			}
+		}
+		s, err := json.Marshal(expected.Schema)
+		if err != nil {
+			return responseError{
+				errMessage: append(errMsgs, fmt.Sprintf("problem marshaling the user provided json-schema: %s", err)),
+			}
+		}
+		schema, err := gojsonschema.NewSchema(gojsonschema.NewBytesLoader(s))
+		if err != nil {
+			return responseError{
+				errMessage: append(errMsgs, fmt.Sprintf("problem generating json-schema schema: %s", err)),
+			}
+		}
+		result, err := schema.Validate(gojsonschema.NewBytesLoader(bodyBytes))
+		if err != nil {
+			return responseError{
+				errMessage: append(errMsgs, fmt.Sprintf("problem validating the body: %s", err)),
+			}
+		}
+		if !result.Valid() {
+			return responseError{
+				errMessage: append(errMsgs, fmt.Sprintf("the result is not valid: %s", result.Errors())),
+			}
+		}
+	} else if expected.Body != "" {
+		if !reflect.DeepEqual(body, expected.Body) {
+			errMsgs = append(errMsgs, fmt.Sprintf("unexpected body.\n\t\thave: %v\n\t\twant: %v", body, expected.Body))
+		}
 	}
 	if len(errMsgs) == 0 {
 		return nil
@@ -362,7 +419,7 @@ func newRequest(in Input) (*http.Request, error) {
 
 func readSpecs(dirPath string) (map[string][]byte, error) {
 	data := map[string][]byte{}
-	files, err := ioutil.ReadDir(dirPath)
+	files, err := os.ReadDir(dirPath)
 	if err != nil {
 		return data, err
 	}
@@ -371,7 +428,7 @@ func readSpecs(dirPath string) (map[string][]byte, error) {
 		if !strings.HasSuffix(file.Name(), ".json") {
 			continue
 		}
-		content, err := ioutil.ReadFile(path.Join(dirPath, file.Name()))
+		content, err := os.ReadFile(path.Join(dirPath, file.Name()))
 		if err != nil {
 			return data, err
 		}
@@ -414,7 +471,7 @@ func (krakendCmdBuilder) getEnviron(cfg *Config) []string {
 	return environ
 }
 
-var defaultBackendBuilder mockBackendBuilder
+var DefaultBackendBuilder mockBackendBuilder
 
 type mockBackendBuilder struct{}
 
@@ -428,7 +485,7 @@ func (mockBackendBuilder) New(cfg *Config) http.Server {
 	mux.HandleFunc("/redirect/", checkXForwardedFor(http.HandlerFunc(redirectEndpoint)))
 	mux.HandleFunc("/jwk/symmetric", http.HandlerFunc(symmetricJWKEndpoint))
 
-	return http.Server{
+	return http.Server{ // skipcq: GO-S2112
 		Addr:    fmt.Sprintf(":%v", cfg.getBackendPort()),
 		Handler: mux,
 	}
