@@ -3,6 +3,7 @@ package krakend
 import (
 	"fmt"
 
+	otellura "github.com/krakend/krakend-otel/lura"
 	cel "github.com/krakendio/krakend-cel/v2"
 	jsonschema "github.com/krakendio/krakend-jsonschema/v2"
 	lua "github.com/krakendio/krakend-lua/v2/proxy"
@@ -13,8 +14,9 @@ import (
 	"github.com/luraproject/lura/v2/proxy"
 )
 
-// NewProxyFactory returns a new ProxyFactory wrapping the injected BackendFactory with the default proxy stack and a metrics collector
-func NewProxyFactory(logger logging.Logger, backendFactory proxy.BackendFactory, metricCollector *metrics.Metrics) proxy.Factory {
+func internalNewProxyFactory(logger logging.Logger, backendFactory proxy.BackendFactory,
+	metricCollector *metrics.Metrics) proxy.Factory {
+
 	proxyFactory := proxy.NewDefaultFactory(backendFactory, logger)
 	proxyFactory = proxy.NewShadowFactory(proxyFactory)
 	proxyFactory = jsonschema.ProxyFactory(logger, proxyFactory)
@@ -22,6 +24,13 @@ func NewProxyFactory(logger logging.Logger, backendFactory proxy.BackendFactory,
 	proxyFactory = lua.ProxyFactory(logger, proxyFactory)
 	proxyFactory = metricCollector.ProxyFactory("pipe", proxyFactory)
 	proxyFactory = opencensus.ProxyFactory(proxyFactory)
+	proxyFactory = otellura.ProxyFactory(proxyFactory)
+	return proxyFactory
+}
+
+// NewProxyFactory returns a new ProxyFactory wrapping the injected BackendFactory with the default proxy stack and a metrics collector
+func NewProxyFactory(logger logging.Logger, backendFactory proxy.BackendFactory, metricCollector *metrics.Metrics) proxy.Factory {
+	proxyFactory := internalNewProxyFactory(logger, backendFactory, metricCollector)
 
 	return proxy.FactoryFunc(func(cfg *config.EndpointConfig) (proxy.Proxy, error) {
 		logger.Debug(fmt.Sprintf("[ENDPOINT: %s] Building the proxy pipe", cfg.Endpoint))
